backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
langchain
langchain-google-genai
google-generativeai
python-dotenv
